Fix bug when loading >32 structures without PLM_USE_SS_IMAGE_VEC; plastimatch can now read both ss_img types without recompiling; additional values shown for plastimatch header; add regression case for 33 structures

+++ src/plastimatch/rtss_structure.h
@@ -32,7 +32,7 @@ public:
     Pstring color;
     int id;                    /* Used for import/export (must be >= 1) */
     int bit;                   /* Used for ss-img (-1 for no bit) */
-    int num_contours;
+    size_t num_contours;
     Rtss_polyline** pslist;
 public:
     plastimatch1_EXPORT
